Update style.css
This commit adds all necessary CSS for the services on the services to expand and collapse when clicked. It also includes some styling changes like the background as well as a section for a photo to be added on the homepage. All in all these changes took like 30 minutes to plan out, try out, and implement.